+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Add support for custom installations
+- Add `queryPipe`
 
 ## [0.2.0] - 2022-06-20