+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Generate types on build
+
 ## [0.3.0] - 2022-06-21
 
 - Add support for custom installations